Jump button now restricts camera roll when zooming
I have came across situations where I would want to take a completely straight zoomed-in screenshot with the camera yet zooming also affected the roll. This adjustment will allow users to hold the jump button to prevent the camera from changing the roll angle.

I thought this way would keep the familiarity of the controls everyone already knows, but I also believe it might be better to make it only roll when holding the jump button.