Update control.Servers.js
What that does is if the gamemode ends with a number it strips out that number. Most TTT servers are "terrortown" so if we're "terrortown2" it'll strip that and group us with "terrortown". It only looks at the last character so "example123" would become "example12" and "example5a" would remain "example5a".
This is the topic where Darth Tealc requested a feature to achieve this end goal: http://facepunch.com/showthread.php?t=1262115