TTT: disarm C4 when teleported
Maps such as 67th way allow their map-based teleporters to be exploited
by using them to teleport C4. Though this is strictly speaking a map
issue, we can detect the general case of C4 suddenly moving from point A
to point B, which is always a problem regardless of how it is achieved.
So disarm the bomb if that happens.