Update table.lua
Adds table.Shuffle( table, times ) to the table extension. This randomizes a table from 1 to many times. It modifies the actual table passed in.

This was created as using "function( a, b ) local _rand = math.Rand( 0, 1 ); _rand > 0.5; end" no longer works as listed in the old wiki. The error is invalid sort method ( although at times it did randomize the table strangely enough ).