TTT: New attempt at propspec fix
Seems there were still problematic cases created by 6aaed736 (see #610),
so this replaces it with a different fix.

The original issue was that players who propspec during prep remained
linked to that entity if they became spectator later during the round.
Rather than trying to clear that linkage for all players when the round
starts, we now clear it in a safer way whenever a player spawns.