TTT: fixed colour issues if not all players have same model
If the random model for a round was a colourable model, any players with
their own custom special rainbow seahorse models would also get a colour
applied.

Fixed by applying colour to a player only if their specific model is
allowed to have it. To prevent overrides of PlayerSetModel from
overriding this colouring behaviour, it has been moved to a
TTTPlayerSetColor hook, which is always called after PlayerSetModel.