Fix Spawnmenu Save button not showing
This also causes a bug in the spawn menu where the save button doesn't show up. To recreate:
- Create two spawnlists, into one of them move any spawn icon
- Click on the save button
- Move the item between the two spawnlists

The save button doesn't appear now, but it should as the list changed. 

This happens because the check in DropAction_Normal in line 88 passes, which calls DropAction_Simple. In DropAction_Simple the OnModify call isn't present, while in DropAction_Normal it is (l. 142)