TTT: Moved ttt_spectate call from InitPostEntity to Initialize.
A disheartening moment for some players is when they've joined just before their traitor round and they are in spectator mode from last time they were on. ttt_spectate is only set in InitPostEntity and this potentially could be after the round has started and the traitors have been picked. An annoyance for both the player and their traitor buddies.

Moving the ttt_spectate call from InitPostEntity to Initialize prevents this.