TTT: Removed role reset from ttt_spectate.
Resetting the role in ttt_spectate only caused troubles. This would throw damage logs off if the player, for example, threw an incendiary grenade or planted a C4 before entering spectator mode. Servers with a defibrillator would cause the player to respawn as an innocent serverside, but still a traitor clientside.