updated jQuery/AngularJS
These are the latest versions of Angular and jQuery that don't break the
main menu.
They improve performace/stability and provide bugfixes.

The Angular version is still seriously outdated though (2012), so we
should consider moving to a newer version if we are going to keep a
"web" driven menu.

Also jQuery-ui wasn't using a minified version, so I replaced it with a
more recent minified one.