Entity:HookOutput( ), Entity:UnhookOutput( )
Creates Entity:HookOutput( name, uid, func ) and Entity:UnhookOutput( name, uid ) to catch entity outputs, especially engine ones. Conforms to the behavior of hooks, where passing the entity as the uid causes it to be passed to the function.

Could obviously use some improvement, but it is currently the only vanilla way to do this that I know.