TTT: Fixed Decoy's radar color
The client was looking for a "-1" when it was actually reading a "3" from the network message. When the networking functions were updated it's likely that this was broken and then overlooked as a "-1" is actually sent by the server but not received as intended.

Not the most thorough edit but now the radar will properly display decoys on the radar to traitors just as before.