TTT: Fixed Rifle scope, zoom spread, unscope delay
There were two gaps when zoomed in with the rifle on the top and bottom of the screen respectively. Fixed by drawing two lines on the top and bottom of the screen.

It was also possible to miss a well aimed shot with the sniper rifle while scoped previously, though it would be a rare occurrence surely. Fixed by removing all spread while scoped.

In the default weapon base, when the weapon is fired it will delay both its primary and secondary attack by the value of the primary attack. This is just fine for weapons that fire relatively quickly but having to wait a second and a half each time you fire the rifle to unzoom seems like an oversight instead of a conscious design choice especially considering the fact that you can switch weapons freely while this is happening. Feels very clunky. Gave it an 0.1 secondary fire delay after shooting. Should be friendly with weapon base updates/customizations too.

Fixed some tabbing inconsistencies.