TTT: Fixed C4 beeps not playing through some walls
I assume the reason this happened before is the culling done to prevent unnecessary network information from being sent. Unfortunately, hearing something on the other side of a wall that's about to kill you is necessary in my opinion. It's seems that this isn't an intended feature as there are many times that the C4 is well out of reach but still within your PVS.

To test the difference plant a C4 in that infamous spot underground between the two houses on dm_richland. Before you wouldn't be able to hear it but now you will.

I'm not very familiar with using ints/uints in networking so if someone has advice to make this more efficient I'd appreciate it. Already did a bit of optimization which should be enough.