TTT: Weapon base adds view punch to bullet dir
Bullets fired did not previously consider the view punch added from firing the weapon, meaning that the center of the cone did not line up with the camera. This was fairly noticeable on weapons with a lot of recoil and/or a fast fire rate.

This edit better reflects the direction that the player is looking.