TTT: Allow detectives to reequip their dropped hat
A vital, long awaited solution that should finally solve this horse shit bug. How are we supposed to get immersed if we can't even pick our own hats back up? Who thought that was a good idea? No wonder TTT is so unpopular.

Comes with a cvar that is enabled by default. I'd send the player a text notification but I'm not sure I want to give translators any more hassle.

A slight sound is made when equipping the hat which only detectives may do if they are missing one for whatever reason. It doesn't have to be your own hat.