Added 4 new functions
I always thought it was odd that the math library didn't include many basic functions such as ways to find the average, mean and mode of a group of numbers. I was originally going to only make these functions work with a variable number of arguments that were only numbers, but then I thought of how hard it'd be to not only get all your numbers together, but to separate them by commas so the function actually worked. Most would just use a table, so that's why I made these functions work with both numbers and tables. The function math.GetNumbersFromTbl would make more sense in the table library, but since it'd be pretty pointless adding additional functions to the table library such as GetBoolsFromTbl to make that function seem fit, I decided to keep it in the math library just because it has something to do with numbers. The reason these functions don't check subtables for numbers is because it'd get a whole lot more complicated when you could just loop through the keys of the table or specify the key containing a group of tables you may want.