Autorun lua_run entity's code if spawnflag is set
This edit would run lua_run entities code when its created if its spawnflag is set. This will eliminate the need to trigger the entity with a logic_auto when all you want to do is run a piece of code once when the map loads. 
I'm running it in the KeyValue hook so it should get called before anything is created, instead after entities are spawned.

Also, does Garry's mod have an official fgd for hammer? I know there are some floating around but I feel that one should be included .
I could look into fixing one up and merge it if that is okay