Fixed PR #1007
In addition to spacing changes, fixed a case where OldHeight would be
overridden to an incorrect value if SetExpanded was called while the
animation was going ( i.e. while the height of the panel was not what it
should be for self.OldHeight purpose )

Added IsValid() checks for self.Contents, and fixed an animation bug
when in spawnmenu opening the tool categories after first opening the
spawnemenu with saved closed categories would make the animation expand
to a wrong value.

I still hate this panel very much, and none of these patches with #1007
and this one make it perfect.