Watery Leeches
A HL2 entity missing from the Half-Life 2 entity tab.
The entity that gets spawned by trigger_waterydeath, recreated with lua
to be able to include it in Garry's Mods Half-life 2 tab with spawn
icon. (Remember the time when the inner pod was missing from Gmod9+?
Yeah. This is another one that's missing.)