Fix c-side error which hangs server on beacon pickup.
I may actually be the only person who bothered to test this thing out :P

When picking up a beacon that has been placed, the server will hang at least once per map for up to a few seconds and will always throw an ambiguous blue "[ERROR]" message, I have identified the cause as being incorrect usage of table.Add, this function very much prefers to have the second argument as the desired index, when inserting tables into number-indexed tables. Providing only the table to be inserted, confuses the hell out of things for reasons that I'm not certain of - a better fix may actually be to alter the default behavior of table.Add, I won't make a pull for that though.