Fix Anti-AFK Exploit
Fixes an exploit to bypass the AFK switcher. The exploit is to use the alias console command to overwrite the "ttt_spectator_mode" console command. This fixes it because (I don't actually know 100% how this works, I don't know the engine of the game, I'm basically just making an educated guess of why this works) toggle treats ttt_spectator_mode as a ConVar, which it is, whereas just calling ttt_spectator_mode 1 treats it as a concommand, and the alias exploit overwrites the concommand. If needed I can provide the exploit but I'm sure you can find it out if you're on Github.