Added ammo accessors and convar support
Adds accessors for all ammo keys, as well as adding support for custom keys to be added. Also, DMG_SNIPER and the ammo flags enums were added since they were missing.

This solves https://github.com/Facepunch/garrysmod-requests/issues/857, https://github.com/Facepunch/garrysmod-requests/issues/646 / https://github.com/Facepunch/garrysmod-requests/issues/861

"This is messy and should be done in the engine." I agree, but this is still a viable solution as the maxcarry accessor was added over a year ago and that was it. Side benefit is that these being implemented in Lua is slightly faster, although it's pretty negligible.