Improve server browser
Weight all servers bellow 30ms the same because there is very little difference in ping.

Stop servers with only a couple people from filling up/appearing on the top of the ranking.

Scatter the server list a little bit by adding a random 1-15ms ping value. This gives a higher variance of different regions. A common complain is "There's all French or there's all German servers". This is because the algorithm treats these evenly.