Add constructor arguments to PANEL:Init
Currently there is no way to initialize a panel with some data, even though there are panels which rely on external data as a dependency. This leaves us having to create our own function just to initialize the panel, and we have to call that function after the panel was already supposed to be initialized through ```PANEL:Init```. Isn't that the whole purpose of ```PANEL:Init```? To initialize the panel and prepare it for consumption? Take an inventory system, as an example.

You have an item panel, which shows an item in your inventory. This item panel has a label and a spawn icon. Obviously this item panel relies on an actual item to show anything useful. An item panel without an item doesn't make sense. So you have an item's data that you perhaps got from a database. You pass that data to the panel, which gives the panel the necessary information to display that item.

```lua
function PANEL:Init(itemData)
	self.Label = self:Add("DLabel")
	self.SpawnIcon = self:Add("SpawnIcon")

	self.Label:SetText(itemData.Name)
	self.Label:SetColor(itemData.Color)

	self.SpawnIcon:SetModel(itemData.Model)
end

vgui.Register("MyItemPanel", PANEL, "Panel")
```

Then when you want to create the panel.

```lua
local panel = vgui.Create("MyItemPanel", nil, nil, item)
```

The two functions I've edited already take varargs, they just don't pass them to ```PANEL:Init```.