Pre/Post Init Scripts
Probably it would be better to implement this on C side, but this is needed.  This would be useful for overriding functions, with a guarantee that the function hasn't been called before by different addons.  Before gmod13, we used to have `lua/includes/enum/` folder that did that exact thing, that is, execute scripts in `includes/enum/` before `autorun/` scripts. Right now, multiple anti cheats are using `init.lua` and other developers can't use that file, for any other purposes, without breaking anticheats.