Remove MarkupObject:Create() (#1542)
This function shouldn't exist - it's a class function of the MarkupObject that... creates a MarkupObject. Calling it on an existing MarkupObject creates an object that now has the metatable of the previous MarkupObject which is improper and inconsistent behaviour for a class function. Even the single use of it calls the function on the MarkupObject metatable as if it were an object. It also updates the __index of the metatable to itself every time needlessly instead of setting it once outright.

This will break no compatibility with previous code since the MarkupObject metatable isn't available outside of a single file, and calling it on an actual MarkupObject results in bad behaviour.