Added custom PANEL:SetPos(x,y) function
The PANEL:SetPos(x,y) doesn't check if the DFrame is within screen boundaries by normal means. That was allowing setting position of the DFrame way out of bounds of the screen, making the player unable to close/move the modal window since its title bar was completely unreachable. My version of the function will actually check if the DFrame will be within screen boundaries when using SetPos function. If it isn't, it will close the DFrame, which would be otherwise out of bounds and throw the error showing the destination position of the DFrame in the error message.