Remove GetWeapons and HasWeapon overrides
GetWeapons: altering the function to return nothing for other players isn't safe as if a player already has clientside Lua access, they can just use debug.getupvalue to get the original version. If not for that, they could also still use GetActiveWeapon, HasWeapon, and the savetable to check for another player's weapons.

HasWeapon: it's shared in the base game now.