Balance Changes
M16 Nerfs
The M16 seems strong; I wanted to bring its DPS and TTK in line with other automatics, while keeping its ability to kill with a headshot and bodyshot. It may still be too powerful, but testing is required..
❌ Delay increased from .15 to .17 (400 RPM to 353 RPM)
❌ Damage reduced from 30 to 28 (Still 4 shots to kill)
✔ Headshot multiplier increased from 2.5 to 2.8 (Still 1 headshot + 1 bodyshot to kill)
	200 DPS to 164 DPS
	.60 seconds to kill to .68 seconds (bodyshots)

MAC10 Buffs
The MAC10 has been very popular as the most accurate automatic; now that the M16 has taken that crown I expect pick rate to go down, but I still want it to be effective.
✔ Delay reduced from .08 to .07 (750 RPM to 857 RPM)
✔ Damage increased from 12 to 13 (9 shots to kill to 8)
	150 to 185 DPS
	.72 seconds to kill to .56 seconds (bodyshots)

M249 Buffs
Not to leave out the M249, a slight RoF buff.
 ✔ Delay reduced from .09 to .08 (666 RPM to 750 RPM)
 	155 DPS to 175 DPS
 	.72 seconds to kill to .64 seconds (bodyshots)

Shotgun Buffs
In preliminary testing, the shotgun felt frustrating and inconsistent. Decreasing spread should help.
✔ Spread cone reduced from .115 to .09