Balance Changes (5/27/2020)
Pistols:
Pistol balance is all over the place right now.
The Glock is currently the best pistol by far. It's way too strong (140 DPS) and too accurate.
❌ Reduced Glock damage from 14 to 12 (140 DPS to 120 DPS)
❌ Greatly increased Glock spread, from .028 to .065 (now matches the MAC-10)

The Desert Eagle is the only pistol that can 1-shot headshot. It's intended to be a high risk, high reward weapon, but right now it's practically useless going for bodyshots (66.6 DPS).
✔ Increased Deagle firerate from 100 RPM to 150 RPM (66.6 DPS to 100 DPS)

The Five-Seven is barely used despite having the highest DPS of the three (160 DPS), good accuracy, and a high rate of fire. I think it's strong but doesn't need changes right now.
⏱ Wait and see if usage increases after the Glock nerf

Primary Weapons:
I think the M16 is in a decent place, but players felt underwhelmed by it. I'm making a few changes to improve the feel of the weapon; I don't want to increase firerate or damage output because I don't want it to outperform the MAC-10 at close range.
✔ Increased M16 magazine size from 10 to 20
✔ Decreased M16 recoil from 1.8 to 1.6
➖ Replaced M16 ironsight zoom  with a 50% spread, 50% recoil reduction effect

Equip Shop Items:
The UMP Prototype is still a no-brainer pick for Detectives, largely because it's so frustrating to fight someone equipped with it. I'd much rather see Detectives buy gadgets for detecting. I'm nerfing the damage to hell to make it less usable in a 1v1 fight; more of a support weapon to use when working with another Detective or a confirmed Innocent.
❌ Removed UMP headshot multiplier (3.5x damage multiplier to 2x)
❌ Reduced UMP damage from 16 to 8 (133 DPS to 66 DPS)
❌ Reduced maximum hitpunch degrees from 5 to 2

The Silenced USP is almost never picked by Traitors. It's difficult to compete with the reliability of the instakill knife, or the potential of C4 and the Roundabomb. In the past, we found giving it one-shot headshots was too powerful. As an experiment, I'm modifying the Shotgun code that gives it an increased headshot multiplier at close range.
✔ Increased Silenced USP damage from 25 to 28 (167 DPS to 187 DPS)
✔ Silenced USP now 1-shot headshots at 10 feet or less, with headshot multiplier falling off from 4x to 2x (112 damage to 56 damage) between 10 and 20 feet (see graph: https://i.imgur.com/uEtG3pJ.png)