Distinguish Brazilian Portuguese and European Portuguese
There are two mainly types of the Portuguese language. There is the Brazilian Portuguese spoken in Brazil, and European Portuguese spoken in Portugal.
Since this file is heavily translated into Brazilian Portuguese, I think it should have a separation between the two types, so if any European Portuguese person wants to create a new translation, he/she will be free to create it.