[TTT] Change "Português brasileiro" to "Português (Brasil)"
Most websites on the internet, when they distinguish the Brazilian Portuguese and European Portuguese they will separate in two ways: the European Portuguese will be described only as "Português" (in English: `Portuguese`) or "Português de Portugal" (in English: `Portuguese from Portugal`) , and the Brazilian Portuguese will be described as "Português brasileiro" (in English: `Brazilian Portuguese`) or "Português do Brasil" (in English: `Portuguese from Brazil`). Or they will separete in another way, that is "Português (Brasil)" (in English: `Portuguese (Brazil)`) for Brazilian Portuguese and "Portuguese (Portugal)" (in English: `Portuguese (Portugal)`) for European Portuguese.

So, to follow the pattern that most websites adopt, here is my PR to change that, since I think it is better to follow the default that most websites use.