Balance Changes (12/4/2021)
Primary Weapons:
The M249 is always fluctuating between too weak and too strong. Right now it's a little weak. Time-to-kill at close range seems fine, so I'm going to increase accuracy a little, trying to improve that at midrange.
✔ Decreased M249 spread from 0.08 to 0.07

The M1014 (Shotgun) feels good right now, but it may be a little too effective at midrange and in fights against groups. To encourage users to close in (or switch to a sidearm), I'm going to reduce ammo capacity rather than change spread.
❌ Reduced clip size from 8 shells to 4
❌ Reduced maximum ammo from 24 shells to 12

The M16 still feels a little underwhelming, even after making it full-auto again. We need to be careful, because a hitscan automatic with good accuracy and recoil will quickly become a death laser.
✔ Increased M16 firerate from 350 RPM to 430 RPM (165 DPS to 200 DPS)

Equip Shop Items:
The UMP Prototype continues to be too good. I want it to be a low-damage weapon that can discombobulate a traitor so teammates can fight them, rather than a way to win every 1-on-1 fight.
❌ Reduced damage from 8 to 6 (67 DPS to 50 DPS)

C4 is fun, and people are starting to engage with it more (going for defuses). Since we play in haste mode (shorter round timers) I'd like to encourage riskier plays and/or area denial by setting the timer for longer. As a reminder, the longer the timer is set for, the fewer 'safe' wires there are to cut when defusing.
✔ Increased defuse difficulty for longer C4 times
    1 safe wire  - 4 minutes to 3 minutes
    2 safe wires - 3 minutes to 2 minutes 15 seconds
    3 safe wires - 2 minutes to 1 minute 30 seconds
    4 safe wires - 1 minute to 45 seconds