TTT: fix weapons disappearing during round reset
As per #1887, weapon entities that are part of the map (as opposed to rearm scripts) are recreated by the map reset. At that point players are holding nothing and could pick up the recreated weapon entity. This would then be stripped again when the players got respawned a moment later.

We avoid this by tracking whether a map reset is in progress and preventing weapon pickups during that time. (Based on a different solution proposed in #1887).

Closes #1887.