New way to move eyes, click and drag panel
When selecting the eyes of an entity by right clicking it is possible to:<br>
Set the position from where you are looking through a panel; <br>
Where it is also possible, I left a squint, in or out. <br>
 <br>
Did not understand? See the  <br>
Demo video:  <br>

https://user-images.githubusercontent.com/28203966/197897650-18053b35-84e9-4527-974b-8068f7f0bc71.mp4

The "TOOL:RightClick" function has been changed to "TOOL:Reload" <br>
Because the new function will be in "TOOL:RightClick":