Toolgun bug fixes (#1739)
I noticed that there is a bug where the toolgun does not call Deploy on the first tool that is selected (both realms, so it's not a prediction thing), it's a simple fix. The bug is caused by SWEP:GetToolObject() being called before SWEP:Think(), which sets SWEP.Mode.

Also fixed Holster being called twice in the name of releasing ghost entities.

And finally, I have also re-enabled ToolObj:Allowed() on the client as FCVAR_REPLICATE now works for Lua ConVars.