Refactor weapon_fists
Refactored for extendability and prediction. This is also using a new fists model that uses activities instead of raw sequences (no idea why Max did it this way) and supports combo punching with both fists. Since the model is packed in the VPK, conflicts shouldn't be an issue. I will attach it when it's complete.