Buff Roundabomb
The Roundabomb has been hard to dial in. It needs to be useful against a group of enemies, but if it's too good it becomes a must-pick in large games.

Reviewing [Valve's documentation](https://developer.valvesoftware.com/wiki/Env_explosion) I found a way to tune it a little more finely. We can adjust explosion radius independent of magnitude (damage).

I think radius is in Hammer units (16 units -> 1 foot), but the docs don't specify. We'll find out testing it on Friday.

✔ Increased radius from 300 to 400 (19 feet to 25 feet)
📝 Magnitude remains at 300
📝 C4 has a magnitude of 200 and radius of 1000 (63 feet)