TTT: Fix C4/Radio sounds not playing outside of PAS (#2203)
Replace use of `sound.Play` for C4 sounds, because it relies on `CPASAttenuationFilter` internally which means players not in the same PAS as the C4 could not hear the sounds, even if in close proximity. The attenuation aspect of the filter has been reimplemented in Lua, so that we retain the functionality of only clients within audible range receiving the sound.