Fix map thumbnail freeze
The reason "Start New Game" would freeze is because Gmod makes requests for 'asset://mapimage/' images for every single map even though the images shouldn't be getting requested for maps that are not being rendered. lazy loading wasn't working.
Instead of returning a dummy image I've swapped it to a CSS background-image. This causes the game to only request thumbnails that are actually on-screen. In theory this also means if you've got 1000 maps in a category but only 50 fit on screen before scroll, it'll only request 50 and gradually request the rest as you scroll, instead of requesting all 1000 from the category.

Additionally changed the category ng-repeat from using a ng-hide/ng-show combo to using ng-if. When using ng-hide/ng-show, Angular repeatedly processes the contents of categories that are hidden. With ng-if, it only processes the active category and its maps.