Improve map thumbnail handling
The reason "Start New Game" would freeze is because Gmod makes requests for 'asset://mapimage/' images for every single map even though the images shouldn't be getting requested for maps that are not being rendered. lazy loading to "asset://" appears to be BROKEN in CEF 86. It works in CEF 137.

So to fix it for now, I've changed the category ng-repeat from using a ng-hide/ng-show combo to using ng-if. When using ng-hide/ng-show, Angular repeatedly processes the contents of categories that are hidden. With ng-if, it only processes the active category and its maps. Therefore it will only request the map images of the current category.