Singleplayer level transitions
OnActivate() would receive a 'transition' boolean if the load is happening due to a Source Engine level transition (trigger_changelevel / changelevel2 etc), similar to in player_manager.OnPlayerSpawn. Multiplayer will always end up being false as Source Engine level transitions are only supported in singleplayer.

If OnActivate 'transition' is true, it sets the loading page to "about:blank" and draws the panel with the last rendered frame prior to load, provided by engine.GetLastFrame()

engine.GetLastFrame() would only exist in menu state, and would return a material of the last game frame that was rendered prior to load, excluding hud/vgui/etc.

The result of this is when a singleplayer game initiates a level transition, instead of seeing the default Gmod loading screen, they see a freeze frame of their game with the "loading" source panel in the bottom-right corner, similar to in HL2.